add: add check for input validity

+++ shop.cpp
@@ -2,56 +2,66 @@
 #include "shop.h"
 
 int main(void) {
+	char *input;
 	int command;
+	bool cont = true;
 
 	printf("--- Menu ---\n");
 	printf("LIST OF COMMANDS\n");
-	
 
 	vector<Item> items;
 
 	do {
 		printf("Enter command: ");
-		scanf("%d", &command);
+		scanf("%s", input);
 
-		// TODO: check inpurt
+		if (sscanf(input, "%d", &command) != 1) {
+			printf("Input must containt command (number)\n");
+			continue;
+		}
 
 		switch (command) {
-			case 1: 
-				add_item(items);
+			case 0:
+				cont = false;
+				break;
+			case 1:
+				printf("list of commands\n");
 				break;
 			case 2: 
-				buy_item(items);
+				add_item(items);
 				break;
 			case 3: 
-				change_item_price(items);
+				buy_item(items);
 				break;
 			case 4: 
-				delete_item(items);
+				change_item_price(items);
 				break;
 			case 5: 
-				discount(items);
+				delete_item(items);
 				break;
 			case 6: 
-				find_item_by_name(items);
+				discount(items);
 				break;
 			case 7: 
-				find_item_with_max_price(items);
+				find_item_by_name(items);
 				break;
 			case 8: 
-				list_items(items);
+				find_item_with_max_price(items);
 				break;
 			case 9: 
-				list_under_100(items);
+				list_items(items);
 				break;
 			case 10: 
+				list_under_100(items);
+				break;
+			case 11: 
 				total_price(items);
 				break;
 			default: 
 				printf("Wrong command. Try again.\n");
 				break;
 		} 
-	} while (1);
+	} while (cont);
 
 	return 0;
 }
